Extract DTO mapping logic
